requirements.txt
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
numpy==1.23.5
pandas==2.2.3
matplotlib==3.9.4
seaborn==0.13.2
scikit-learn==1.6.1
Pillow==11.2.1
gdown==5.2.0
tensorboard==2.10.1
h5py==3.13.0
scipy==1.13.1
tqdm==4.67.1 